requirements.txt
# Google Cloud
google-cloud-storage==2.5.0
db_dtypes==1.1.1
google-cloud-bigquery==3.0.0

# Web Scraping
beautifulsoup4==4.11.1
requests==2.28.1

# Data Processing
numpy==1.23.5   # Stable version known to work
pandas==1.5.3   # Compatible with numpy 1.23
openpyxl==3.0.10

# Workflow
apache-airflow==2.5.1